pyproject.toml
[tool.poetry]
name = "storm"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.22.0"
python-telegram-bot = "^12.2.0"
beautifulsoup4 = "^4.8.2"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
